Refactor project structure and remove unused files
