+++ dotnet/api/scripts/set-templates-path.sh
@@ -1,29 +1,31 @@
 #!/bin/bash
 
+# BASE -> API
+export PATH_TEMPLATES_BASE_CONFIG_DOTNETTOOLS="$TEMPLATES_PATH/base/.config/dotnet-tools.json"
+export PATH_TEMPLATES_BASE_VSCODE_LAUNCH="$TEMPLATES_PATH/base/.vscode/launch.json"
+export PATH_TEMPLATES_BASE_VSCODE_TASKS="$TEMPLATES_PATH/base/.vscode/tasks.json"
+export PATH_TEMPLATES_BASE_API_CONTROLLERS_SHARED_APICONTROLLERBASE = "$TEMPLATES_PATH/base/Api/Controllers/Shared/ApiControllerBase.cs"
+export PATH_TEMPLATES_BASE_API_CONTROLLERS_V1_EXAMPLECONTROLLER = "$TEMPLATES_PATH/base/Api/Controllers/v1/ExampleController.cs"
+export PATH_TEMPLATES_BASE_API_EXTENSIONS_APIVERSIONINGSETUP = "$TEMPLATES_PATH/base/Api/Extensions/ApiVersioningSetup.cs"
+export PATH_TEMPLATES_BASE_API_EXTENSIONS_SWAGGERSETUP = "$TEMPLATES_PATH/base/Api/Extensions/SwaggerSetup.cs"
+export PATH_TEMPLATES_BASE_API_PROPERTIES_LAUNCHSETTINGS = "$TEMPLATES_PATH/base/Api/Properties/launchSettings.json"
+export PATH_TEMPLATES_BASE_API_APPSETTINGSDEVELOPMENT = "$TEMPLATES_PATH/base/Api/appsettings.Development.json"
+export PATH_TEMPLATES_BASE_API_DOCKERFILE = "$TEMPLATES_PATH/base/Api/Dockerfile"
+export PATH_TEMPLATES_BASE_API_DOCKERFILELIVE = "$TEMPLATES_PATH/base/Api/Dockerfile.live"
 
-# Define path for the class templates
-export CONTROLLER_BASE_TEMPLATE_PATH="$TEMPLATES_PATH/class/ApiControllerBase.cs"
-export API_VERSIONING_TEMPLATE_PATH="$TEMPLATES_PATH/class/ApiVersioningSetup.cs"
-export DBCONTEXT_TEMPLATE_PATH="$TEMPLATES_PATH/class/ApplicationDbContext.cs"
-export NATIVE_INJECTOR_TEMPLATE_PATH="$TEMPLATES_PATH/class/NativeInjectorConfig.cs"
-export PROGRAM_TEMPLATE_PATH="$TEMPLATES_PATH/class/NativeInjectorConfig.cs"
-export SWAGGER_TEMPLATE_PATH="$TEMPLATES_PATH/class/SwaggerSetup.cs"
+# BASE -> APPLICATION
+export PATH_TEMPLATES_BASE_APPLICATION_INTERFACES_IEXAMPLESERVICE = "$TEMPLATES_PATH/base/Application/Interfaces/IExampleService.cs"
+export PATH_TEMPLATES_BASE_APPLICATION_MAPPINGS_EXAMPLEMAPPINGPROFILE = "$TEMPLATES_PATH/base/Application/Mappings/ExampleMappingProfile.cs"
+export PATH_TEMPLATES_BASE_APPLICATION_SERVICES_EXAMPLESERVICE = "$TEMPLATES_PATH/base/Application/Services/ExampleService.cs"
+export PATH_TEMPLATES_BASE_APPLICATION_VIEWMODELS_EXAMPLE_EXAMPLEVIEWMODEL = "$TEMPLATES_PATH/base/Application/ViewModels/Example/ExampleViewModel.cs"
+export PATH_TEMPLATES_BASE_APPLICATION_VIEWMODELS_PAGEDRESULT="$TEMPLATES_PATH/base/Application/ViewModels/PagedResult.cs"
 
-print_info "Controller Base Template Path: $CONTROLLER_BASE_TEMPLATE_PATH"
-print_info "Api Versioning Template Path: $API_VERSIONING_TEMPLATE_PATH"
-print_info "DbContext Template Path: $DBCONTEXT_TEMPLATE_PATH"
-print_info "Native Injector Template Path: $NATIVE_INJECTOR_TEMPLATE_PATH"
-print_info "Program Template Path: $PROGRAM_TEMPLATE_PATH"
-print_info "Swagger Template Path: $SWAGGER_TEMPLATE_PATH"
+# BASE -> DOMAIN
+export PATH_TEMPLATES_BASE_DOMAIN_COMMON_BASEENTITY = "$TEMPLATES_PATH/base/Domain/Common/BaseEntity.cs"
+export PATH_TEMPLATES_BASE_DOMAIN_ENTITIES_EXAMPLE = "$TEMPLATES_PATH/base/Domain/Entities/Example.cs"
+export PATH_TEMPLATES_BASE_DOMAIN_ENUMS_EXAMPLEENUM = "$TEMPLATES_PATH/base/Domain/Enums/ExampleEnum.cs"
+export PATH_TEMPLATES_BASE_DOMAIN_INTERFACES_IEXAMPLEREPOSITORY = "$TEMPLATES_PATH/base/Domain/Interfaces/IExampleRepository.cs"
+export PATH_TEMPLATES_BASE_DOMAIN_INTERFACES_IBASEREPOSITORY = "$TEMPLATES_PATH/base/Domain/Interfaces/IBaseRepository.cs"
+export PATH_TEMPLATES_BASE_DOMAIN_VALIDATIONS_DOMAINEXCEPTIONVALIDATION="$TEMPLATES_PATH/base/Domain/Validations/DomainExceptionValidation.cs"
 
-
-# Define paths for the Docker-related templates
-export DOCKERFILE_TEMPLATE_PATH="$TEMPLATES_PATH/docker/Dockerfile"
-export DOCKERFILE_LIVE_TEMPLATE_PATH="$TEMPLATES_PATH/docker/Dockerfile.live"
-export DOCKER_COMPOSE_TEMPLATE_PATH="$TEMPLATES_PATH/docker/docker-compose.yml"
-
-print_info "Dockerfile Template Path: $DOCKERFILE_TEMPLATE_PATH"
-print_info "Dockerfile.live Template Path: $DOCKERFILE_LIVE_TEMPLATE_PATH"
-print_info "docker-compose Template Path: $DOCKER_COMPOSE_TEMPLATE_PATH"
-
-
+# BASE -> INFRASTRUCTURE
