+++ dotnet/api/templates/base/.config/dotnet-tools.json
