+++ dotnet/api/templates/base/.vscode/launch.json
