+++ dotnet/api/templates/base/.vscode/tasks.json
