+++ dotnet/api/templates/base/Api/Controllers/Shared/ApiControllerBase.cs
