+++ dotnet/api/templates/base/Api/Controllers/v1/ExampleController.cs
