+++ dotnet/api/templates/base/Api/Dockerfile.live
