+++ dotnet/api/templates/base/Api/Extensions/ApiVersioningSetup.cs
