+++ dotnet/api/templates/base/Api/Extensions/SwaggerSetup.cs
