+++ dotnet/api/templates/base/Api/Properties/launchSettings.json
