+++ dotnet/api/templates/base/Api/appsettings.Development.json
