+++ dotnet/api/templates/base/Application/Interfaces/IExampleService.cs
