+++ dotnet/api/templates/base/Application/Mappings/ExampleMappingProfile.cs
@@ -1,0 +1,15 @@
+using AutoMapper;
+using {{PROJECT_NAME}}.Application.ViewModels.Example;
+using {{PROJECT_NAME}}.Domain.Entities;
+
+namespace {{PROJECT_NAME}}.Application.Mappings
+{
+    public class ExampleMappingProfile : Profile
+{
+    public ExampleMappingProfile()
+    {
+        CreateMap<ExampleViewModel, Example>();
+        CreateMap<Example, ExampleViewModel>();
+    }
+}
+}
