+++ dotnet/api/templates/base/Application/Services/ExampleService.cs
