+++ dotnet/api/templates/base/Application/ViewModels/Example/ExampleViewModel.cs
