+++ dotnet/api/templates/base/Application/ViewModels/PagedResult.cs
