+++ dotnet/api/templates/base/Domain/Common/BaseEntity.cs
