+++ dotnet/api/templates/base/Domain/Entities/Example.cs
