+++ dotnet/api/templates/base/Domain/Enums/ExampleEnum.cs
