+++ dotnet/api/templates/base/Domain/Interfaces/IBaseRepository.cs
