+++ dotnet/api/templates/base/Domain/Interfaces/IExempleRepository.cs
