+++ dotnet/api/templates/base/Domain/Validations/DomainExceptionValidation.cs
