+++ dotnet/api/templates/base/Infra.Data/EntitiesConfiguration/ExampleConfiguration.cs
