+++ dotnet/api/templates/base/Infra.Data/Repositories/BaseRepository.cs
