+++ dotnet/api/templates/base/Infra.Data/Repositories/ExampleRepository.cs
