+++ dotnet/api/templates/base/docker-compose.yml
