--- dotnet/api/templates/class/Infra.IoC/NativeInjectorConfig.cs
@@ -1,28 +1,0 @@
-using {{PROJECT_NAME}}.Application.Interfaces;
-using {{PROJECT_NAME}}.Application.Mappings;
-using {{PROJECT_NAME}}.Application.Services;
-using {{PROJECT_NAME}}.Domain.Interfaces;
-using {{PROJECT_NAME}}.Infra.Data.Context;
-using {{PROJECT_NAME}}.Infra.Data.Repositories;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-
-namespace {{PROJECT_NAME}}.Infra.IoC
-{
-    public static class NativeInjectorConfig
-    {
-        public static void RegisterServices(this IServiceCollection services, IConfiguration configuration)
-        {
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseNpgsql(configuration.GetConnectionString("DefaultConnection")));
-
-            services.AddScoped<IExampleService, ExampleService>();
-            services.AddScoped<IExampleRepository, ExampleRepository>();
-
-            services.AddAutoMapper(typeof(ExampleMapper));
-
-        }
-    }
-}
